Updated to use JDK 8
In order for Bamboo versions newer than 5.9.x to run in this container, you need to use JDK 8.  This is because in Bamboo 5.10 Atlassian switched to tomcat 8.  Failing to use the newer JDK causes the launch to die with "unable to start listeners".  Similar changes may be necessary for the bamboo-agent image, but I haven't gotten that far yet.